Rename plural
